// react/src/features/admin-space/product/product-type.ts
export type ProductPayload = {
  name: string;
  description: string;
  category_id: number;
  unit_price: number;
  stock_qty: number;
  image_link: string;
};

export type ProductType = {
  id: number;
  name: string;
  description: string;
  category_id: number;
  unit_price: number;
  stock_qty: number;
  image_link: string;
};